Tons more indexer runtime optimizations
Lots included here. Ideally these changes would be committed
individiually; future changes will be committed more orderly
once the core implementation is fully implemented and stable.

Overall, many of the changes pertain to NNG event handling
and the associated logic, i.e. stable reorg handling

- Bugfix: `syncBlocks()` is now passed the checkpoint `Block`

Previous logic would result in an outdated checkpoint state
during runtime. `syncBlocks()` was adjusted to work with the
checkpoint `Block` object and return that checkpoint when
sync is completed, rather than returning `null`.

- Adjust timing of `QUEUE_STATE` toggle

Previous logic was blocking event triggers from adding their
callbacks to the event loop. This was to avoid desynchronizing
the runtime state and avoid exceptions.

The timing of this toggle was suboptimal at best, and was removed
from activating on the `blkconnected` trigger. Stress testing
with hundreds of RANK txs per second caused no discrepancies
with the runtime state, nor caused any exceptions or race
conditions.

- Subscribe to NNG `blkdisconctd`; implement handler

This was previously not implemented. Having this in place now
was critical to properly handling chain reorgs.

The aforementioned `QUEUE_STATE` toggle was moved to this
function (and may be removed when `txmempoolrm` is implemented).

All tests using reorgs via `invalidateblock` calls worked with
no issues.

- Set `REPEATABLE READ` on SQL transactions

The database queries are structured to spend as little time
as possible in nodejs. The query data is structured in node,
then passed to the db function to execute as a single SQL
transaction.

Prisma's default setting for transactions is `RepeatCommited`,
which adds unnecessary overhead. The indexer is not concerned
with the state of the rest of the database during execution.

Relations between tables are few, and Profile rankings can be
incremented and decremented in any order, so long as all RANK
values are applied successfully.

- Add `try/catch` blocks to many functions

Proper error handling and reporting (duh)

- Several miscellaneous changes

Renamed some functions, moved some functions, fixed up log strings,
and many other minor optimizations to runtime execution (e.g.
reinstated a mempool RAM cache to avoid SQL queries for reconciling
block txs missing from the database)