bugfix: block rewind only processing one block
The `rewindBlocks()` method was designed to rewind a range of blocks.
However, a previous commit moved the logging to the calling method, rather
than to log the results of each rewind.

This commit moves the block range to the calling method. `rewindBlocks()`
is also changed to only rewind a single block. This way the calling
method controls the block range and logs the results in the appropriate
context.