v0.0.1 - Massive architecture overhaul
This commit introduces a huge refactoring of the runtime logic. It also
introduces some minor but impactful changes to the database schema and
SQL statements. This code has undergone a battery of tests and has
performed rock solid. Runtime logic is much cleaner, extremely more
performance, and less prone to exceptions. With this commit, this branch
is able to process 9,000+ RANK tx/s on initial sync.

Summary of major changes:

- Add `mempooltxrem` handler

This functionality was critical to properly handling block reorg events.
Conflicting RANK txs are now seamlessly rewound from the database along
with their associated profiles.

- Reinstate mempool cache

The mempool cache was reinstated following the addition of `mempooltxrem`.
Keeping a cache of RANK txs in memory allows upsert/rewind logic to
work much more seamlessly and drastically reduces I/O load on the
database.

- Significant database performance improvements

The adjustments made to the database schema provided the avenue for
drastically simplify the SQL statements and move them into SQL
transactions. Every SQL query that is now executed (minus the ones that
operate on a single table/record) are now executed in atomic fashion.
The SQL statements also take full advantage of the database relations.

Other noteworthy changes:

- Tons of minor (and major) tweaks/optimizations

- Completely refactor console logging scheme

- Optimize types

- Redo many comments and JSDocs