v0.0.1 - Indexer core functionality (#1)
* Add .gitignore

* Adjust .gitignore

* Initial commit of `indexer-ts`

This is the indexer that will parse Lotus transactions for specific
protocol data and store it in a PostreSQL database.

* Add `bn.js` submodule

* Implement the core indexer functionality

So many changes but here's a brief summary:
- Database module is fully integrated
- Block sync is fully integrated (AND FAST)
- Mempool sync is fully integrated
- NNG subscriptions and RPC calls are fully integrated
- Runtime is relatively stable under heavy load
- Significant (ongoing) architectural adjustments
- Several module adjustments (all moved to sub_modules/ folder)

Things not yet completed by this commit:
- Reorg detection and handling is in progress
- Streamline setup
- Provide documentation
- Misc. checks and balances

* Tons more indexer runtime optimizations

Lots included here. Ideally these changes would be committed
individiually; future changes will be committed more orderly
once the core implementation is fully implemented and stable.

Overall, many of the changes pertain to NNG event handling
and the associated logic, i.e. stable reorg handling

- Bugfix: `syncBlocks()` is now passed the checkpoint `Block`

Previous logic would result in an outdated checkpoint state
during runtime. `syncBlocks()` was adjusted to work with the
checkpoint `Block` object and return that checkpoint when
sync is completed, rather than returning `null`.

- Adjust timing of `QUEUE_STATE` toggle

Previous logic was blocking event triggers from adding their
callbacks to the event loop. This was to avoid desynchronizing
the runtime state and avoid exceptions.

The timing of this toggle was suboptimal at best, and was removed
from activating on the `blkconnected` trigger. Stress testing
with hundreds of RANK txs per second caused no discrepancies
with the runtime state, nor caused any exceptions or race
conditions.

- Subscribe to NNG `blkdisconctd`; implement handler

This was previously not implemented. Having this in place now
was critical to properly handling chain reorgs.

The aforementioned `QUEUE_STATE` toggle was moved to this
function (and may be removed when `txmempoolrm` is implemented).

All tests using reorgs via `invalidateblock` calls worked with
no issues.

- Set `REPEATABLE READ` on SQL transactions

The database queries are structured to spend as little time
as possible in nodejs. The query data is structured in node,
then passed to the db function to execute as a single SQL
transaction.

Prisma's default setting for transactions is `RepeatCommited`,
which adds unnecessary overhead. The indexer is not concerned
with the state of the rest of the database during execution.

Relations between tables are few, and Profile rankings can be
incremented and decremented in any order, so long as all RANK
values are applied successfully.

- Add `try/catch` blocks to many functions

Proper error handling and reporting (duh)

- Several miscellaneous changes

Renamed some functions, moved some functions, fixed up log strings,
and many other minor optimizations to runtime execution (e.g.
reinstated a mempool RAM cache to avoid SQL queries for reconciling
block txs missing from the database)

* bugfix: block rewind only processing one block

The `rewindBlocks()` method was designed to rewind a range of blocks.
However, a previous commit moved the logging to the calling method, rather
than to log the results of each rewind.

This commit moves the block range to the calling method. `rewindBlocks()`
is also changed to only rewind a single block. This way the calling
method controls the block range and logs the results in the appropriate
context.

* Add event queue, adjust reorg logic, fixes/tweaks

- Fully implemented queue system

Previous logic would execute the callback for each NNG message event.
This would cause issues with messages being processed out of order,
leading to some events cascading the callback triggers and crashing
the node. This commit fully implements a queue system, whereby NNG
messages that are received are pushed into a `pending` queue and
resolved in the order they are received.

Adding the queue system also fixed a race condition that was being
worked around with an in-memeory `mempool` array.

- Adjust `blkdisconctd` handler to per-block

Previous logic would attempt to reconcile an entire reorg within a
single NNG event trigger. This necessitated adding other workarounds
that overcomplicated things.

With the queue system in place, those workarounds were no longer
needed, as each disconnected block will now be processed in correct
order without issues.

- Return calls to `this.close()` when exception is thrown

Previous logic would allow runtime to continue after an exception
was thrown and the indexer was shutting down. This commit prepends
a `return` to each call to `this.close()` to ensure smooth shutdown.

- Database schema optimizations

Optimized schema and corrected some relations

- Function refactoring/renaming and misc fixes/tweaks

Tons of small optimizations and semantics adjustments

* bugfix: `txid` corruption

A bug was introduced that corrupted the `txid` while processing tx
outputs. This commit reverts back to a `for()` loop to keep the
`Transaction.Output[]` array intact, thus preventing corrupting the
'txid' property.

* Validate RANK script chunks

Previously, we were only checking for the presence of the RANK LOKAD
prefix. This commit adds logic to validate all of the required RANK
script chunk values. This ensures that malformed output scripts do not
cause runtime exceptions.

* v0.0.1 - Massive architecture overhaul

This commit introduces a huge refactoring of the runtime logic. It also
introduces some minor but impactful changes to the database schema and
SQL statements. This code has undergone a battery of tests and has
performed rock solid. Runtime logic is much cleaner, extremely more
performance, and less prone to exceptions. With this commit, this branch
is able to process 9,000+ RANK tx/s on initial sync.

Summary of major changes:

- Add `mempooltxrem` handler

This functionality was critical to properly handling block reorg events.
Conflicting RANK txs are now seamlessly rewound from the database along
with their associated profiles.

- Reinstate mempool cache

The mempool cache was reinstated following the addition of `mempooltxrem`.
Keeping a cache of RANK txs in memory allows upsert/rewind logic to
work much more seamlessly and drastically reduces I/O load on the
database.

- Significant database performance improvements

The adjustments made to the database schema provided the avenue for
drastically simplify the SQL statements and move them into SQL
transactions. Every SQL query that is now executed (minus the ones that
operate on a single table/record) are now executed in atomic fashion.
The SQL statements also take full advantage of the database relations.

Other noteworthy changes:

- Tons of minor (and major) tweaks/optimizations

- Completely refactor console logging scheme

- Optimize types

- Redo many comments and JSDocs