Fixes and cleanup (#2)
* Migrate from old repository, dir change

* Significant package cleanup, adjust `.gitignore`

* `npm i @types/express@^4` and `npm audit fix`

* bugfix: db missing profile ranking/rank counts

A bug was introduced that removed the deconstructed `Profile` object
from the profile database upserts. This was preventing the calculated
`ranking`, `ranksPositive`, and `ranksNegative` properties from being
upserted into the appropriate database fields. This commit fixes that,
and also adjusts the schema so these fields are `NOT NULL`.

* Add `prettier` + rc/ignore

* `npm i typescript-eslint` and add `eslint.config.mjs`

* Update `.gitignore` to keep `local_modules/`

We are using a slightly modified version of the actively-maintained
nanomsg package. Our version adjusts the npm `install` script to use
the globally-installed version of `node-gyp` to work around a syntax
error. Have not found an ideal solution to this yet.

* Schema field name changes

`value` => `sats`
`ranksPositive/ranksNegative` => `votesPositive/votesNegative`

* Prettier formatting + minor tweaks

Renamed the `getScriptPartHex` method to `getScriptPart` and added
an optional `encoding` parameter to return the `ScriptPart` bytes in the
specified encoding.

Also updated/removed some comments and whatnot. This commit is primarily
for Prettier formatting.

* Update package details + `rank-indexer-ts@0.0.1`

Many of the details in `package.json` were outdated since migrating
from the old repository. Those have been updated and the package version
was set to 0.0.1 to revert in correct package version set in previous
commit.

* Mitigate vulnerabilities + add `.npmrc`

Running `npm audit fix` revealed several severe vulnerabilities, many
related to the local `nanomsg` package. This commit adds an `.npmrc`
to the project to omit the installation of `devDependencies` by default.
The `rank-indexer-ts` package dependenices were also adjusted to
ensure the availability of required packages and modules for runtime.