More fixes and cleanup (#8)
* bugfix: `sentiment` always positive

Previous behavior was performing a falsey conditional on a non-empty
string, which would always be true. This commit properly checks the
`sentiment` value to calculate accurate profile rankings.

* Convert `OP_RETURN` hex data to UTF-8 for database

Previous behavior would insert the raw hex data into the database. This
presented unnecessary complexity with the API module. This commit
encodes the raw hex data into UTF-8 before inserting into the database.

* indexer.ts: prettier formatting

* Remove `express` + dependencies

This is moved to `rank-api-ts`.

* Add `@lotusia/rank-suite` as parent module

This commit adds the `@lotusia/rank-suite` package and sets the local
path accordingly. Several components that were part of this package are
being moved to the `rank-suite`.

* database.ts: get account vs profile by `profileId`

The `rank-api-ts` package will be fetching the entire account (i.e. all
profiles associated with the human identity) and then filter to the
appropriate platforms where necessary.

* prettier