Fixes and structure (#9)
* Migrate types/imports to `@lotusia/rank-suite`

* Simplify script chunk processing

Previous behavior of using `enum`s and reverting to/from hex/UTF8 which
imposed unnecessary complexity. This commit changes RANK script parsing
logic to use the new script chunk `Map`s (from `@lotusia/rank-suite`)
for simplicity.

* Simplify tx output processing

Previous behavior was designed to accommodate multiple RANK outputs per
tx, however the spec does not allow for this. This commit fixes the
output processing logic such that only a single RANK output can be
parsed from any single tx. Preparation is also made for RANK extension
outputs (i.e. on-chain comments)

* Reduce NNG queue function calls

Previous behavior would call `nngProcessMessage()` before returning from
`nngReceiveMessage()`, which would then check if the queue was busy.
This commit moves the queue/busy check to the `nngReceiveMessage()` call
so that we are not calling and returning and wasting CPU cycles for no
reason.

* Comment + log entry adjustment

Cosmetics

* Move `app.ts` to `scripts/start.ts`

* Prisma: Connect Block/RankTransaction `timestamp`

Previous behavior would not update the timestamp for the unconfirmed
`RankTransaction` database entries. This commit adjusts the schema so
that the relation between the Block and the `RankTransaction[]` includes
the `timestamp` field.

* Rename module

Moved the module to the package root as `index.ts` + adjust import

* Rename `Indexer` module, move to `lib/`

The `Indexer` module is a class of the overall backend.

* indexer.ts: rename `parts` to `chunks`, misc changes

Rename the RANK script `parts` class property to align with types and
other methods. Also make some other minor misc. adjustments.

* Add constants, types, functions + adjust imports

These were previously moved to the `rank-suite` repository, but this
design decision was reverted.

* Remove `rank-suite` + adjust packages, tsconfig

Remaining changes to unify the RANK backend services.

* Adjust NNG module imports

For some reason these modules were hairpinning back into their
directory to import other NNG modules.

* Remove NNG TCP support

It's not necessary.

* blkdisconctd: don't include `prevhash` in `Block`

Adding the `prevhash` is not necessary for runtime processing and adds
unnecessary bloat to the log entry.

* Database.ts: deconstruct `Profile` for processing

* Update `.prettierignore`

* Update `tsconfig.json`

* Update `README.md`