Rework `RankOutput` processing logic
This commit reworks the entire `RankOutput` processing logic so that
the Bitcore script `Chunk` objects are used to validate the output
script. This required a significant adjustment to the core types and
constants.

Also, the defined optional script chunks are now processed and added
to the `RankOutput` object if found. These optional chunks are
platform-dependent, so the hard-coded platform parameters are now used
for this processing.