Enforce minimum satoshi value in RANK output
Previously, there was no check to validate that Lotus had even been
burned to rank a profile or post. This would not have caused any issues
with the profile/post rankings, however it could have allowed an
attacker to very cheaply flood the indexer database with empty
RANK outputs.

This commit hard-codes and enforces a minimum output value of 1 XPI
(1_000_000 satoshis). This may still be too low and will likely be
increased in the future.