Implement new protocol spec (#10)
* Rework `RankOutput` processing logic

This commit reworks the entire `RankOutput` processing logic so that
the Bitcore script `Chunk` objects are used to validate the output
script. This required a significant adjustment to the core types and
constants.

Also, the defined optional script chunks are now processed and added
to the `RankOutput` object if found. These optional chunks are
platform-dependent, so the hard-coded platform parameters are now used
for this processing.

* Enforce minimum satoshi value in RANK output

Previously, there was no check to validate that Lotus had even been
burned to rank a profile or post. This would not have caused any issues
with the profile/post rankings, however it could have allowed an
attacker to very cheaply flood the indexer database with empty
RANK outputs.

This commit hard-codes and enforces a minimum output value of 1 XPI
(1_000_000 satoshis). This may still be too low and will likely be
increased in the future.

* Add `postId` database model and processing logic

This is a significant change to the RANK protocol. Previously, there
were definitions for the `Post` script chunks, but they were never
processed and indexed.

This commit adds all of the appropriate runtime and database logic to
completely implement `Post` rankings. The `Post` model has been defined
in the Prisma schema and has relations to both the `Profile` and
`RankTransaction` models. If a RANK tx has a `postId` defined, the
`RankTransaction` is linked to the `Post` model that is inserted into
the database. The `Post` also appropriately adjusts the ranking of the
associated `Profile` model. Full upsert and rewind logic is tested and
working.

* whoops: remove the invalid type import

This constant was removed in 21649a0db622a35b403b806bccc65e45d33c7d76.
The import is now removed.