Add REST API endpoint (#11)
* Add API module + database support

This commit adds the initial `lib/api.ts` library and implements the
required database functionality to support it.

* constants.ts: Add `API_SERVER_PORT`

Hard-code HTTP server listening port

* Use new `Database` instance

The API was originally configured to utilize an existing instantiated
`Database` for its queries. This commit makes the API set up its own
`Database` instance, and also sets up the `init()` and `close()` methods
to connect and disconnect the database respectively.

* Manage HTTP server via `init()`/`close()`

The parent module will no longer be expected to manage the HTTP server
of the API. Instead, the API module itself will intiialize and tear down
the HTTP server during `init()` and `close()`.

* Fix `postId` parameter typo

`postid` should be `postId`

* Enable `postId` param handler

* Add detail to 404 responses

* `GET` -> `get`

It's cleaner this way.

* Add types for param and get handlers

Should have been this way in the beginning. :)

* Add `/stats` endpoints, adjust Prisma indexes

This commit adds API endpoints for profiles/posts that will be useful
later for activity tracking. The Prisma schema indexes for `Profile` and
`Post` models was also adjusted so that these new API calls will be as
performant as possible.

* Define `top-ranked`/`lowest-ranked` result count

Database queries were set with a raw integer, but this commit adds a
defined constant for this purpose.

* Set `PlatformParameters` once as global settings

Previously the API was applying `this.app.set()` upon every API call,
but this is unnecessary. This commit sets the `platformParams` config
upon instantiation and is referenced throughout runtime.