Include profile data in `postId` API request (#15)
* Return profile data in `postId` request

This will be useful in reducing the amount of necessary API calls made
by `rank-extension-ts`.

* `Database`: Include profile in return object

Neglected to ensure that the included `profile` data was being returned
to the API after executing the database query. This commit fixes that.

* Return profile ranking as string

JSON cannot serialize a bigint; need to convert profile ranking to string
first.

* Always return data, even if post doesn't exist

This commit changes the logic to execute an interactive transaction that
will first fetch a post, and if that doesn't exist, fetch the profile.

If neither exist, a default data object is returned with 0 ranking,
0 positive votes, and 0 negative votes.

* Remove unneeded log line

* Rework interactive transaction logic

Wrap the calls in try/catch/finally and revert the queries to
`findUniqueOrThrow`. This way the post query fails, then falls back to
the profile query, which finally falls back to returning the default
data object with any changes that had been made while executing the
transaction query.