Workaround: "Attempt to access memory outside buffer bounds" (#17)
* Add `converter` method to platform `postId` params

Having only a `reader` method for `PlatformParameters` was insufficient,
as there are edge cases that require an alternative means of converting
a hex value into original encoding, rather than relying only on Buffer
readers.

For example, old Twitter posts were formatted differently than they are
now. Some old post IDs may have been BigInt values, but were not 64-bit
unsigned integers. Therefore, using the `readBigUInt64BE` reader failed,
as the BigInt did not fill the required byte length.

This commit adds a secondary `converter` method to the `PlatformParameters`,
which will be used as a fallback in case the primary `reader` method fails.

* Ignore failed `reader` calls

For now, we will just fallback to ranking the profile rather than the
post if the `postId` fails to process with the associated platform
`reader` function.

This will be fixed later.