Fix `postId` param reader validation (#19)
* Ignore `postId` param reader validation

* Twitter: enforce `postId` regex start/end

Previously we were just making sure that there was at least 1 number in
the string, and this is not correct. This commit enforces the start and
end boundaries to ensure that the `postId` is what we expect for Twitter.

* Twitter: re-enable `bigint` parsing for `postId`

With fixing the regex test, this can safely be re-enabled, as only numbers
will be processed in the `BigInt` function.

* Send `application/json` MIME type for GET requests

Previous was set to `application/javascript` which is depcretated; most
if not all clients will expect `application/json`, so we make that change
here.