Add schema, indexer, and API integration for `scriptPayload` (#23)
* Update `rank-lib` submodule

* Add `scriptPayload` field, index to Prisma schema

* `database.ts`: accept `scriptPayload` for API query

This commit adjusts the `apiGetPlatformProfilePost` API query to accept
a `scriptPayload` string as an optional parameter. If specified, the
RANK transactions associated with the `scriptPayload` will be returned
as `postMeta` (metadata), along with the ranking for the post and profile.

* `indexer.ts`: Add `scriptPayload` indexing

This commit adds a new method to the `Indexer` class that processes
the inputs of a `Transaction` to ensure the same address spent UTXOs to
pay for the RANK transaction. If there is more than one P2PKH that spent
UTXOs, the RANK transaction will not be processed.

* `api.ts`: Add `scriptPayload` parameter

This commit adds processing logic for the `scriptPayload` GET parameter.

* `indexer.ts`: remove `comment` chunk processing

We will no longer be needing the `comment` processing, as this data chunk
will be moved to the additional `OP_RETURN` outputs.