Add timespan parameter to `/stats` API queries (#25)
* Add timespan parameter to `/stats` API queries

This commit adds the requisite conditional to database queries for fetching
top- and lowest-ranked profiles/posts over the specified timespan. The
API was also appropriately adjusted to accommodate the new `timespan`
GET parameter.

* Include txids in `/stats` API response

This commit adds the `votes` string array to the `/stats` API responses
for both profiles and posts. This will be useful for dashboard display,
showing viewers the on-chain proofs that aggregate into the profile/post
rankings.

* Add `includeVotes` bool param for including txids

When making calls for long timespans, there will be many txids to return.
This commit adds the `includeVotes` conditional to avoid always including
txids in API responses.

* Change default timespan to `day`

* Add pagination for `includeVotes`

This commit is added to prevent DoS vector. Default page size is 10 votes.

* Change other `/stats` methods to `day` timespan

Forgot to include the other database queries in 6ac38b0

* Add `pageNum` string to `Parameter` type

This should have been included in 01a69b7