Adjust API and IPC database queries (#36)
* Add `today` to `Timespan` type

This will allow us to set an end to the `Timespan` range on API/IPC database
queries. This makes queries to the `/stats` API and IPC endpoints deterministic.

* `Database`: Make IPC queries deterministic

This commit changes the function parameter definitions into an object that
defines a `startTime` and `endTime`. Database queries are now using this
`endTime` parameter to get a deterministic range of results.

* Fix `/stats/posts/*` API queries, make deterministic

This commit simultaneously makes the same changes to the function
definition parameters as in 0ad8b61. This commit also fixes the initial
`groupBy` query to not fetch rows that have a null `profileId` or `postId`.

Some of the logic was also streamlined to reduce conditionals.