bugfix: restore `finally` block to API queries (#39)
Previously these were removed due to satisfy eslint rule `no-unsafe-finally`,
which prevents a `return` (among other terminating statement) in the
`finally` block. However, this introduced undesired behavior, whereby the
`prisma.Profile.findUniqueOrThrow()` query would not return data, resulting
in HTTP 404 errors on the front-end (as the API module was not receiving
data to respond to the client request).

This commit reinstates the `finally` blocks in these queries and disables
the eslint rule in the entire file.