Implement extension registration API endpoint, configure Temporal
This branch implements all of the required pieces for the API to accept
registration data from extension installations. As part of the Lotusia
onboarding process, new extension installations will automatically
opt-in to registration with the back-end. This registration provides
initial funding for the wallet, and in the future will be necessary for
engagement rewards, receiving news/updates from Mrs. Turtle, etc.

Temporal dependencies were also installed, as these will be used to
execute automations in the backend systems.