Add JSDoc and optimize runtime processing
This branch adds JSDoc across the runtime modules. The NNG processing
queue of the indexer is optimized to process messages in batches. The
Database module now converts `ProfileMap` and `PostMap` objects to async
iterators when converting them to Prisma statements. The API now does
not require the `profile` parameter when accessing the `/stats`
endpoint.