Implement multiple OP_RETURN outputs (#68)
* Adjust Prisma schema for multiple outIdx

With adding neutral votes to the same tx, we need to track output indexes for the RANK outputs. This commit adds a new field to the `RankTransaction` model with a default `0` outIdx to handle backwards compatibility.

* Adjust `indexer.ts` to handle multiple RANK outputs

This commit makes several changes to the indexer module to accommodate multiple OP_RETURN outputs, specifically for `neutral` outputs with 0 value

* Adjust `database.ts` to connect RANK `txid`s with `outIdx`

* Update `rank-lib`