Significant updates to database types, query processing, return object structures
This commit helps unify and make sense of the different types of RANK/RNKC object formats, and where they are relevant. For example, RNKC transactions queried from the database may have relational data attached to them, and these object structures require a type that unifies Prisma relational data with the base RNKC data. Also, API queries need their own data structure because `bigint` values are not serializable to `application/json`.

A minor update to the Prisma schema was also added, which requires a full re-indexing of the RANK database to accommodate.