Include voter details for profile in API by default
This commit adjusts the `apiGetPlatformProfile` database method to include the voter details by default. This is to fix a regression in the Lotus website which now does not load profile pages. Also, there is no other way to specify whether or not to include voter details in these requests, so just include them by default. :)