Update types, add `profileMeta`, misc changes
This commit adjusts the types in the database module to be more specific
and accurate. This commit also adjusts the post API request and database
queries to return a `profileMeta` property in the Post object. This
property will be useful for upcoming changes to the extension.

There are also some other minor type and function changes made here.