Do authCacheEntry expiry check before authorization check (#90)
This commit fixes a procedural logic bug that was deleting entries from the `AuthCache` before the validation logic. This was throwing an exception in the API, causing re-authentication  in stall at "Refreshing auth info..." message.

The expiry check is now executed and expired entries are now deleted prior to auth validation.